+++ lib/src/client/systems/controller_system.g.dart
@@ -6,12 +6,26 @@ part of 'controller_system.dart';
 // SystemGenerator
 // **************************************************************************
 
+abstract class _$KeyboardControllerSystem extends GenericInputHandlingSystem {
+  Mapper<Camera> cameraMapper;
+  SettingsManager settingsManager;
+  _$KeyboardControllerSystem(List<Element> ignoreInputFromElements)
+      : super(Aspect.empty()..allOf([Camera]), ignoreInputFromElements);
+  @override
+  void initialize() {
+    super.initialize();
+    cameraMapper = Mapper<Camera>(world);
+    settingsManager = world.getManager<SettingsManager>();
+  }
+}
+
 abstract class _$ControllerSystem extends EntityProcessingSystem {
   Mapper<Booster> boosterMapper;
   Mapper<Controller> controllerMapper;
   GameStateManager gameStateManager;
   CameraManager cameraManager;
   ControllerManager controllerManager;
+  SettingsManager settingsManager;
   _$ControllerSystem() : super(Aspect.empty()..allOf([Booster, Controller]));
   @override
   void initialize() {
@@ -21,5 +35,6 @@ abstract class _$ControllerSystem extends EntityProcessingSystem {
     gameStateManager = world.getManager<GameStateManager>();
     cameraManager = world.getManager<CameraManager>();
     controllerManager = world.getManager<ControllerManager>();
+    settingsManager = world.getManager<SettingsManager>();
   }
 }